+++ apps/web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @doccov/web
 
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @doccov/sdk@0.25.10
+  - @doccov/api-shared@0.2.13
+
 ## 0.1.14
 
 ### Patch Changes