+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @doccov/sandbox
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @doccov/sdk@0.25.10
+
 ## 0.1.9
 
 ### Patch Changes